+++ CHANGES.rst
@@ -38,6 +38,8 @@ Unreleased
  * **[FEATURE]** Added :meth:`get_multireddit` to get a single multireddit obj
  * **[FEATURE]** Added :meth:`get_my_multis` to get all multireddits owned
    by the logged in user.
+ * **[FEATURE]** Added :meth:`get_multireddit` to :class:`Redditor` to quickly
+   get a multireddit belonging to that user.
 
 PRAW 2.1.19
 -----------